backend/Pythonservice/requirements.txt
flask==3.0.0
flask-restx==1.3.0
flask-cors==4.0.0
werkzeug==3.0.1
google-generativeai==0.3.2
chromadb==0.4.22
numpy<2.0.0
